[mcp-account-manager] Use libsignon to read credentials
The authoritative source for an account's username is in its auth
credentials; manually storing that in the account as well can be
desynchronized and fail.

Instead, read the credentials from signon and set the telepathy account
name based on that. This still does involve mirroring the username to
an account setting, but it can be done in a reasonably automated way.

Currently, this only happens for account creation; it should be changed
in the near future to update as needed.