[telepathy-sasl-signon] Only set CredentialsNeedUpdate flag when appropriate
Previously, the CredentialsNeedUpdate flag would be set in a variety
of situations, including when a generic OperationFailed error occurred.
This commit ensures that it is only set for an account when a
UserInteraction error occurs (that is, the account credentials are
valid but need to be refreshed manually by the user).