[telepathy-sasl-signon] Use g_message to log CredentialsNeedUpdate
Previously, we used g_debug to log the CredentialsNeedUpdate event,
but g_debug output is suppressed by default.  This commit changes
it to use g_message instead, which is not suppressed, to ensure
that the event shows up in the journal logs.

This commit also suppresses some warnings by switching to newer
API to avoid deprecation warnings, and marking some function
parameters as unused.